existential type (*) becomes "any"
Summary:
This diff essentially ends support for the long deprecated existential type
https://flow.org/en/docs/types/utilities/#toc-existential-type

From now on `*` will still be parsed but will be an alias for `any`.

This diff also removes the `codemod replace-existentials` command, since
it relies on logic that is now being retired.

Changelog: [feature] The deprecated `*` type is now an alias to `any`. You can use `flow codemod replace-existentials`, which is available in previous versions of Flow, to upgrade your codebase to not use `*`. This codemod command will not be available in the current or future versions of Flow.

Reviewed By: evanyeung

Differential Revision: D31411440

fbshipit-source-id: cc29805e62fa4a95f96050fc002132c54ec90309