Special case havocing of variables whose only closure writes are providers
Summary:
Flow is currently pretty clever about both when to havoc variables, and what to havoc them to. The first part, when to havoc, is the same with the new environment semantics. The second part--what to havoc to--is different.

The new environment semantics always havocs to the full set of providers for the variable, if the variable is havocable. On the other hand, with the old semantics, when the havocing was because of a function call (as opposed to a function scope entry or a yield), we would only havoc to the type of all writes to the variable that occur *within* a nested function. So, in the following program, we don't havoc `x` to `null` because that write isn't in a nested scope, so the call to `havoc` couldn't reach it.

```
var x = 42;
if (condition) {
  x = null;
}

if (typeof x === 'number') {
  console.log(); // havocs
  (x: string | number); // ok: x is not null
}

function set_x() {
  x = "hi";
}

```

The new environment is pretty much incompatible with this design in general--this cleverness depends entirely on tvars, and on flowing an open tvar (representing the write inside the closure) to the havoced type well before actually seeing the write inside the closure itself.

However, we can recover *some* of this behavior in the special case where the *only* writes to the variable that occur in nested scopes are also *providers* for the variable. This could come up in a case like this:

```
var x = null;
function initialize_x() {
  x = 42;
}

if (typeof x === 'number') {
  console.log(); //havocs
  (x: number); // fine!
}
```
In this case, when we havoc `x` at the call to console.log, we can know that the only write to `x` that can happen as a result is `x = 42`, and since that's a provider for `x`'s type, that write will already be resolved in the environment. So, rather than havocing to the *entire* set of providers for `x`, which include `null`, we only havoc to the providers that are in nested function scopes (unioned with the pre-existing type of the variable).

We lose that ability, though, if a non-provider write exists, because we may not have the resolved RHS type of that write, and we have no way of knowing if the write will reset the variable to the type of a non-nested provider. For example, in the above program, we will have a more general havoc and the cast will now fail if we add the function

```
function reset_x() {
  x = null;
}
```
since now `console.log()` could theoretically call `reset_x()` and lose the refinement.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D31185382

fbshipit-source-id: d0373d3068b55767e985237ec58326c1c850a6c1