Fix handling of assigned-to globals
Summary:
There were a couple issues with globals/undeclared variables in the provider analysis and env builder.

First, the provider analysis originally located undeclared variables in the nearest var scope to the write to the variable. This is incorrect for JS, where undeclared variables live in the root var scope.

Second, the ssa_builder's computation of unbound names only considered variables that are read from. If an unbound name only existed on the LHS, it wasn't included in the result from the ssa_builder, which meant that the env builder didn't know to consider it as a global, which in turn meant that writes to it were ignored in the env builder, which finally means that the loc_env didn't create an entry for the location of the write, which led to crashes in D31121535.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D31132082

fbshipit-source-id: ae91c5a680169f54fc2207af4c4779a6b8bac98a