Change error formatting to handle newlines
Summary:
Currently error messages shouldn't contain newlines, though this is (as far as I can tell) undocumented. If a newline is included, it will be printed, but the formatting code won't realize that it's on a new line and will insert its own newlines when it thinks it has reached the correct number of characters to go to the next line, ignoring the fact that it already has. This means that, if the terminal was 10 characters across and I inserted a message like, `"Error: foo has:\nsome probs"`, it would be formatted like
```
 Error: foo
 has:
 some
 probs
```
instead of

```
 Error: foo
 has:
 some probs
```
because after formatting "some", the formatter thinks it's 9 characters into the line that starts with "has".

This diff changes that by allowing the formatter to recognize inline `\n` characters, split on them, and use them to start new lines that it recognizes, which allows things to print correctly

Changelog: [internal]

Reviewed By: mroch

Differential Revision: D31224789

fbshipit-source-id: 638a020c212aaffbef895911e7dc2b19a18bf301