Use provider analysis to power refinement invalidation in the old environment
Summary:
In addition to rolling out constrained writes D30970514 (https://github.com/facebook/flow/commit/10c17c0d13a31d445224bbcfe534ab86edad4a30), we have the opportunity to implement the new, provider-based behavior for havocing in the old environment as well! In this behavior, for any potentially havocable variable (so vars, lets, classes, and functions), when we `bind` them and determine whether they're actually havocable, we also look up their providers (using the same repurposing of the loc_env as in D30970514 (https://github.com/facebook/flow/commit/10c17c0d13a31d445224bbcfe534ab86edad4a30)), turn them into a single type, and store them in the variable's scope entry. (For variables that aren't havocable or where the provider analysis doesn't apply--consts, imports, internal vars--we set the provider to be the same as the general type.)

Then, when we havoc, we simply set the specific type to the provider type. (If the `ProviderHavoc` environment option is not enabled, the provider type is always exactly the general type, so this part of the implementation is not gated under the config option.)

This diff also has a bugfix for `declare export default class`, which was handled incorrectly in the hoister previously.

Reviewed By: panagosg7

Differential Revision: D31121534

fbshipit-source-id: 44971975ce8725282f0ee5782775d20fc56a1555