disallow comments inside private identifier names
Summary:
comments are not allowed after the `#` in private identifier names. currently we allow `#/*derp*/foo`.

**Note:** Private identifiers are not identifiers! `Flow_ast_mapper` used to recurse into `identifier` on these, which could lead to problems. Now we don't, but _that_ could also change behavior. I'm relying on tests here {emoji:1f62c}

 ---

**Per the spec:**

`PrivateIdentifier` is a `CommonToken` which is an `InputElementDiv`. comments, line terminators and whitespace are only allowed in `InputElementDiv`.

since they're siblings of `CommonToken`, they're not allowed inside `CommonToken`, and so not allowed between the `#` and `IdentifierName` in `PrivateIdentifier` like they might be between other tokens/punctuation.

[InputElementDiv](https://tc39.es/ecma262/#prod-InputElementDiv) ::
- WhiteSpace
- LineTerminator
- Comment
- CommonToken
- DivPunctuator
- RightBracePunctuator

[CommonToken](https://tc39.es/ecma262/#prod-CommonToken) ::
- IdentifierName
- PrivateIdentifier
- Punctuator
- NumericLiteral
- StringLiteral
- Template

[PrivateIdentifier](https://tc39.es/ecma262/#prod-PrivateIdentifier) ::
- `#` IdentifierName

Changelog: [parser] Fix a bug that allowed comments after the `#` in private identifiers

Reviewed By: gkz

Differential Revision: D31423043

fbshipit-source-id: 9b93ae7bb24015c5569e768e2da435a341338c63