fix ensure_parsed with prioritize_dependency_checks
Summary:
during a recheck, we call `ensure_parsed_or_trigger_recheck` to verify that files haven't changed on disk. that can request that more files get rechecked and then calls `raise Lwt.Canceled` to cancel the current recheck.

when `experimental.prioritize_dependency_checks` is enabled, we do a recheck just of prioritized files (unchecked deps) and defer the rest for a subsequent recheck.

the problem is, those extra `ensure_parsed` files weren't prioritized and so we end up in an infinite loop: a prioritized recheck cancels itself to add some `ensure_parsed` files. the next recheck prioritizes the same files but defers the ensure_parsed files, so it cancels again, ad nauseum.

the fix is to track these extra files separately (`files_to_prioritize`) and include them in the next recheck.

Reviewed By: vrama628

Differential Revision: D31506304

fbshipit-source-id: fe2b1312505a77f11a09f9df70f2762e85e44f02