Add use op information for constrained writes
Summary:
This diff changes use_op for our post-inference checks of variable writes against their providers, to include a new frame `ConstrainedAssignment`. This frame currently is ignored for error formatting, but includes information about the variable that was reassigned that may be useful for error messages in the future and that is necessary for codemodding fixes
to these variables.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D31290463

fbshipit-source-id: 2227e6c42de29e7023ee0a570fda6426b73f7d7b