Fix to the ssa builder's handling of for-in/of loops
Summary:
This diff fixes an issue where a variable declaration in a for loop entry was seen to not be a write by the ssa builder. This was because the ssa builder's definition of `variable_declarator` skips declarations with no `init`, but in e.g.
```
for (var x of [1,2,3]) { x }
```
the declarator node for `x` doesn't have an `init` node but is still a real write to `x`.

It turns out the env_builder handles this exactly correctly already, so this diff just essentially copies over the implementation from there.

Changelog: [internal]

Reviewed By: jbrown215

Differential Revision: D31420287

fbshipit-source-id: 80cc93e9e0efce9b7ecb7a11cb6a411c9e30303b