Codemod to annotate variable declarations with constrained write errors
Summary:
This is the first of two codemods that will support switching to writes being constrained by their providers.

This codemod's role is to add annotations to variable declarations that are involved in constrained write errors. Essentially, we find the location of the declaration of any variable with a constrained write error (from the use_op added in the previous diff), and then when we traverse that declaration, we add an annotation based on the general type of the variable (which includes all writes to it, even ones that trigger constrained write errors).

Reviewed By: panagosg7

Differential Revision: D31290462

fbshipit-source-id: 5f9453814b37e52a035a2aa6a5d11d6af87a120c