Less repositioning during merge
Summary:
Note: this change applies for the current merge phase, not just new-merge.

This diff reduces the amount of repositioning done during the merge phase.
Type_sig_merge calls repositioning in a couple of occasions. This operation
creates a new lazy tvar that gets as payload a repositioned type.

Before, we would assign a repositioned reason to the lazy tvar, and also reposition
the type (through `Flow_js.reposition`). With this diff, we only reposition the
reason of the lazy tvar and skip the second repositioning.

The rationale for this change is that when dealing with annotations, like the merge
phase does, we don't need to deal with repositioning as much; the locations
we care about are encoded in the type structures themselves (AnnotT, ClassT,
etc.). The minimal error diff is justification for this argument.

More motivation for this change is that porting the full current reposition
implementation to annotation inference can be challenging. So, instead I decided
to even out this change by disabling this extra repositioning from merge completely.

Reviewed By: jbrown215

Differential Revision: D29422810

fbshipit-source-id: 94994ddb45cf9d5cdef9ccd087be97b635a10c17