Introduce Heap Refinements
Summary:
This diff adds partial support for Heap Refinements to the EnvBuilder. Specifically, we:
1. Start to add refinements to heap entries
2. Havoc heap refinements on function calls

But we do not:
\1. havoc heap refinements when shared projections are written. For example:

```
let x = {};
let y = x;

if (x.foo === 3) {
  y.foo = 4;
  (x.foo: 3); // should fail
}
```
\2. Add heap refinements on writes

```
let x = {};
x.foo = 3;
(x.foo: 3); // should not fail
```

There are also several items related to optional chaining that will need to be addressed in future diffs.

These features will be handled in diffs stacked on this one.

We model heap refinements as a mapping from projections to Val.t's and store it alongside the `val_ref`s already present on env entries.

The primary complexity in this diff lies in the implementation of:
1. `merge_heap_refinements`
2. `merge_env`/`merge_remote_env`/`merge_self_env`
3. `merge_conditional_branches_with_refinements`
4. `scout_changed_vars`
5. `add_refinement`
6. `expression`

This diff also fleshes out the functions that we used to gather refinement keys to also report heap refinement keys. I moved them to their own module as well, and that module is growing increasingly similar to `typing/refinement.ml`, as expected.

Changelog: [none]

Reviewed By: mvitousek

Differential Revision: D30380300

fbshipit-source-id: 68f10fcaec6e38f9c9116355fb7728ac8c999b04