export_binding should query providers, not read
Summary:
This callsite of `var_ref` is asking for the flow-specific type of an exported var. That sounds like the wrong type to export-- instead it should be looking at the general type of the var.

In the new environment, this means we query the providers instead of the EnvBuilder output.

New test results:
`Passed: 169, Failed: 157, Skipped: 268, Errored: 141`

Reviewed By: mvitousek

Differential Revision: D30814150

fbshipit-source-id: 01e9f6ebb8c24446a1873d17eaf99f483fb2e33a