Latent Refinements
Summary:
This diff implements latent refinements. A latent refinement is one introduced by a predicate function. We call them latent because we do not know if a function introduces a refinement until after we have its type.

To support Latent Refinements we introduce a `LatentR` refinement which mirrors the structure fo the `LatentP` `Type.predicate` as closely as possible. Instead of supplying a type we supply the location of the callee, which can be looked up in the environment for its type.

Changelog: [Internal]

Reviewed By: mvitousek

Differential Revision: D29698315

fbshipit-source-id: da390b3b79344b479fddb6ea8950a78a7717bb20