Get rid of global_TODO, remnants from pre-global implementation in EnvBuilder
Summary:
The SsaBuilder reports all unbound names to the EnvBuilder. No need for these SMap.find_opt's anymore!

There was a bug in ssa_builder.ml where unbound variables that are written to but never read were not reported to the EnvBuilder. This diff fixes that bug as well.

Changelog: [none]

Reviewed By: mvitousek

Differential Revision: D30611144

fbshipit-source-id: ce851b21d885c5bf3b0ab2ac17f8615beacaaf9c