fix infinite recheck loop when deleting a file w/ saved state
Summary:
consider the new test in this commit:
- C depends on B which depends on A
- with saved state, we don't parse files until we need to check them, so we haven't parsed B
- B is deleted but the file watcher hasn't seen that yet

previously, when you check `A`:
- we calculate direct dependents and cache it with `DirectDependentFilesCache`
- since B is a direct dependent, we need to check it too.
- we call `ensure_parsed` to parse B so we can check it, but it's missing
- we redo the recheck, now including that B is deleted (same as if we'd gotten the delete FS event)

here we hit the bug: the cache key we pass to `DirectDependentFilesCache` only includes new or changed files, not deleted files. so when we redo the recheck, we hit the cache and still think that `B` is a direct dependent.

so we plan to check `B` and call `ensure_parsed`, but it's still missing, so we redo the recheck again. over and over, forever!

if the cache wasn't there, we'd calculate direct dependents of `{A, B}` and see that C needs to be rechecked as a dependent of B, and _crucially_ that B does _not_ need to be rechecked.

the fix here includes deleted files in the cache key, so that we get a cache miss when we redo the recheck.

Reviewed By: panagosg7

Differential Revision: D31611731

fbshipit-source-id: 53f98a671870395120f4a242c2f09edcb505f00f