Filter out node modules from dependent sets when determining what to check
Summary:
Given files A and B, where B depends on A and both are within a node_modules
directory, forcing A in a lazy server would cause B to be checked.

With a non-lazy server, neither of the above files would be checked with, since
according to a comment on `unfocused_files_to_infer`, we should never be
checking files in node_modules.

The reason for this inconsistency was that in `determine_what_to_recheck`
files in node_modules were not filtered out from `sig_dependent_files` and
`all_dependent_files`, and so `include_dependencies_and_dependents`
would include them in the `to_check` set.

This diff filters out such files from `*_depentent_files` sets. There's probably
other places where this pruning can happen. I chose this cause it's near the call
to `unfocused_files_to_infer` which does the same filtering on the `input` set.

Changelog: [fix] fix bug where errors in node_modules were shown in lazy server

Reviewed By: mroch

Differential Revision: D31600744

fbshipit-source-id: 4544ad398637c55f538d02e9b2bb9e6a96241cf4