During post inference constrained-write checking, treat tvars with no LBs as empty for soundness
Summary:
This diff changes post-inference tvar resolution for the purposes of constrained write checking, to turn tvars with no lower bound into `empty` (rather than `any`). This fixes a number of soundness holes that were likely blockers to rolling out this behavior, but it does have user-visible results: it's now an error to write to providers that contain unresolved, empty tvars. This can occur in a number of places, most notably:
1. Arrays: it's an error to do `let x = []; x = [42]`
2. Functions: it's an error to do `let x = (a: number) => 42; x = b => 42` (because `b` has no lowers)
3. Generics: it's an error to do `let x = new Set(); x = new Set<number>();`

This diff also makes a small change to the actual `new_env.ml` to prevent this issue from over-agressively blocking initializers of unannotated declarations.

Reviewed By: panagosg7

Differential Revision: D31496351

fbshipit-source-id: 5b03debcaf039811fd56ddec6aa4c3dcad7a8a53