Codemod to rename redefined variables with constrained write errors
Summary:
This is the second codemod supporting constrained writes.

This codemod is intended to handle the cases where a variable is written to multiple times, not because the variable is conceptually union-like, but rather that a later assignment essentially reuses the variable name, but is entirely distinct from the original variable. This is cases like

```
var x = 42;
x = x > 42 ? "hello" : "nope";
(x: string)
```
In these cases, it's essentially incorrect to annotate `x` with `number | string`: doing so will introduce errors at the use-site of `x`, since that use expects it to just be a number. It's also unnecessary, since at every point in this program it's clear what type `x` is.

Instead, the right thing to do with this program is to extract the second definition of `x` into a separate variable:
```
var x = 42;
const xCond = x > 42 ? "hello" : "nope";
(xCond: string)
```
This will allow LTI to handle this program without any annotations needing to be added, and will avoid the conflation of `x`'s type that would occur if we just ran the annotation codemod over this program.

This diff introduces a codemod that extracts such variables into separate declarations. This codemod makes two passes over the AST: first, it determines which assignments to variables should be extracted into separate declarations, and then it applies that renaming through the program. Both passes rely heavily on the SSA builder and Scope builder analyses; these tools allow this transformation to be really precise and handle uses of variables before their definitions, havoced variables, etc without issues.

The initial analysis pass of this diff determines whether a particular assignment (call it "A") to a variable "X" based on the following criteria:
1. Was X involved in a constrained write error?
2. For each read of X, does the set of writes that reach the read (according to the SSA builder) either equal exactly A and nothing else, or do not include A?
3. Is A *not* a provider for X?

If all these are true, we're allowed to extract this assignment into a new variable declaration. This means that a case like

```
var x = 42;
x = x > 42 ? "hello" : "nope";
(x: string)
```
above will qualify to be renamed, because every read of `x` either reaches *only* the second assignment, or it doesn't reach the second assignment at all.

On the other hand, a case like
```
var x = 42;
if (condition) {
  x = "a";
  x;
}
x;
```
will not allow a renaming, because the last read of `x` reaches *both* assignments to `x`. If we renamed the second assignment, it would be unclear what would be read from at the end. Likewise, anywhere where there's a havoc will cause the SSA builder to think that any later reads from `x` will reach all writes, and so prevent renaming.

Generally, the way I imagine this being used is that this codemod will initially be run against codebases that are being upgraded, in order to extract out variables where possible, before then running the annotation codemod D31290462 (https://github.com/facebook/flow/commit/cad8bf7227fbf4b983b6a71e03b7e03cb6820f7e) against them afterwards to fix the remainder of the cases by adding annotations.

 ---
This diff also has some heuristics for what to rename an extracted variable to. It tries to be clever in cases like `x = createFooContainer(x)`, where it will rename `x` to `xFoo`. This is a place where additional cleverness may be needed as we roll out these changes.

Reviewed By: panagosg7

Differential Revision: D31453075

fbshipit-source-id: 9c40acd77f8fd675ced9d2019e1773c3d43a4367