Add information in the annotate-declarations codemod about possibly renamable variables
Summary:
The interplay between the annotate-declarations codemod and the rename-redefinitions codemod is such that it might be useful for the former to suggest that the latter be run. This diff adds stats information to the annotate-declarations codemod about how many variables should perhaps be renamed rather than annotated. It's a little imprecise: `var x;` will count towards the number of assignments that should be renamed according to this diff, but won't actually be renamed by the rename-redefinitions codemod.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D31453076

fbshipit-source-id: f8a1c0608788ad8f915caee636c37c78764f9eeb