Type JSON as object rather than a class
Summary:
`JSON` should be typed as an object with two properties rather than a class. We currently allow you to do `new JSON()` which is a runtime error, and prevent method unbinding which should be allowed (reported in #8689).

We also allowed you to use `JSON` as a type (which many people in WWW apparently have done), but that type is meaningless - it is impossible to instantiate a `JSON` instance (though obviously this wasn't the intention of the authors, who probably incorrectly assumed it was the type returned from `JSON.parse` or something similar)

Changelog: [fix] Type `JSON` as object rather than a class

Reviewed By: mroch

Differential Revision: D31588184

fbshipit-source-id: 8674391cd068e92aeb8e90346c6a5c2522655222