Heuristic simplifications for declaration annotation codemod
Summary:
This diff adds an option to attempt to simplify some of the generated annotations for constrained writes.

As of D31496351 (https://github.com/facebook/flow/commit/4be61998815666532507989e09f322b5066a06e2), tvars in with no lower bounds are treated as empty for the purposes of constrained write checks, which causes errors in cases like `let x = []; x = [42]`. When we attempt to run the annotate_declarations codemod to fix these issues, originally the generated type of `x` here would be `Array<number> | Array<empty>`. However, the  `empty` array here is essentially spurious, an artifact of how we perform inference and now how the program is expected to behave (or how it behaved at all until the new environment stuff rolls out).

This diff adds a `--filter-deep-empty` option to the annotate declarations codemod that tries to filter out cases like this. It looks at unions of generated `Ty.t`s and, in cases where all members of the union are arrays, removes those members that contain generated (not annotated) `empty` types as the element type of the array. It does similar for functions and for generics.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D31548641

fbshipit-source-id: 164beeaae9e886ed3df66be58c8775ed1dc19ba0