Fix for providers of function parameters
Summary:
Minor fix: we previously only installed providers when variables are initialized in env.ml (because in almost all cases, that's the point at which we have the type of the variable). However, annotated function parameters are never initialized, only bound, with the type annotation already in hand, with the variable already initialized. This diff handles installing providers in that case.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D31547783

fbshipit-source-id: 4d1cb2fb2ffc34e5ae7a7fbe954bc23bddbd76cf