Fix for unreachable code
Summary:
Originally we assumed that all variables would have providers, but the `env_builder` is responsible for providing the env with the set of keys that are providers, and the env builder skips unreachable code. So if there is a provider within unreachable code, we would discover a variable with no providers.

This is an issue that should be resolvable in the new env, but for the purposes of the old env this diff just treats variables with no providers as unconstrained.

Reviewed By: panagosg7

Differential Revision: D31621631

fbshipit-source-id: b193e7f5f487606e97d6f36fccbea3dad0ce0686