GetPropT in annotation inference
Summary:
* Instantiates GetPropT_helper with primitives implemented within annotation inference.
D31292191 explains how the abstraction around in GetPropT logic is imperfect, since
the respective code handles cases that are irrelevant to annotation inference. This
can be seen in the implementation of `read_prop`: Flow_js relies on `perform_lookup_action`
for this, a function that handles all actions (not just reads). In Annotation_inference, we
are only interested in reads, so `read_prop`s logic is duplicated to reflect that.

* Adds relevant entries in `elab_t` pattern match, that mimic the ones in flow_js.ml

Reviewed By: jbrown215

Differential Revision: D31293478

fbshipit-source-id: d0d43f214c3f0465695dc8bdb7f09536559281c1