Improve error messages for constrained writes and add the ability to include an explanation to errors
Summary:
This diff makes a number of changes in service of having good error messages for constrained write errors.

First, when we see a constrained write error, we want to explain why the error exists. The format we're going for in a program like `var x = "hello"; x = 42` is
```
Cannot assign `42` to `x` because number [1] is incompatible with string [2]. All writes to `x` must be
compatible with the type of its initializer [3]. Add an annotation to `x` [3] if a different type is desired.
```
That is, it's a normal compatibility error, but with the added explanation at the end. It takes some doing to accomplish this, however, because of the way that frame use ops work. If we just made a new root use op that represented constrained assignments, we would produce nonsense if that root use op was wrapped in a frame use op: for example, if we did `var x = ["hello"]; x = [42]` the error we'd get will include a frame use op indicating that this is happening in the array element, and we'd end up with
```
Cannot assign `[42]` to `x` because number [1] is incompatible with string [2]. All writes to `x` must be
compatible with the type of its initializer [3]. Add an annotation to `x` [3] if a different type is desired
in array element.
```
The "in array element" here is misplaced. (Well, I guess it does kind of make sense here, but that's essentially coincidental.) "in array element" is intended to refer to the incompatibility of number with string, but because of the way that frames are layered on around the initial message, it's hard to directly connect them.

We solve this by creating a new kind of component to an error message: an "explanation". This is intended to be a sentence that explains the error but is not a part of the core error message, and which can be appended after the message. So while currently the structure of an error is
```
ROOT in FRAME in FRAME ... in FRAME. [ERROR-CODE]
```
it now becomes
```
ROOT in FRAME in FRAME ... in FRAME. EXPLANATION. EXPLANATION. ... EXPLANATION. [ERROR-CODE]
```
With this capability, the message about why variables are constrained becomes an Explanation and we can correctly output
```
Cannot assign `[42]` to `x` because number [1] is incompatible with string [2] in array element. All writes to
`x` must be compatible with the type of its initializer [3]. Add an annotation to `x` [3] if a different type is
desired.
```
-----
The next change this diff makes is to modify the reasons of any type that is resolved to `empty` during the post-inference constrained write check. We want to indicate that types that are resolved to empty are missing necessary type information, so we indicate textually that Flow lacks the necessary type information. For example, if we've resolved an implicit instantiation to empty in a program like `var x = new Set(); x = new Set<number>()`, we'll format the implicitly instantiated tvar for `new Set()` as "unknown implicit instantiation of `T`" rather than just "`T`".
-----
Finally, to help with cases like `var x = (a: number) => 42; x = (y => y)` where the problem is that `y` receives no lower bounds and is resolved to empty, we'll note that `y` has an unknown type as above, but we'll also suggest that `y` should be annotated as well as/instead of `x`, which we do by recognizing the new `RUnknownParameter` reason desc when formatting the incompatibility error. In this case, the overall error we produce becomes
```
Cannot assign function to `x` because parameter `y` of unknown type [1] is incompatible with number [2] (consider
adding a type annotation to `y` [1]) in the first parameter. All writes to `x` must be compatible with the type of its
initializer [3]. Add an annotation to `x` [3] if a different type is desired. [incompatible-type]
```

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D31589295

fbshipit-source-id: 5790ceb7e58a540798aad64d4aafc086677ce394