When using constrained writes, LHS of assignment should reflect provider types for type-at-pos
Summary:
With constrained writes, the type of the LHS of an assignment/declaration according to type-at-pos should reflect the type of the variable's providers, not the general type of the variable, since the providers specify what assignments to the variable are allowed.

However, the annotate_declarations codemod depends on the general type being the type of the LHS in the typed AST in order to correctly annotate variables by their classic behavior. As such, this diff introduces a new environment mode `ClassicTypeAtPos`, only used programatically by the annotate_declarations codemod, which falls back to the general type for the LHS of assignments.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D31593579

fbshipit-source-id: 5f2052aeed507c041e50e14a70f5452a8048d411