Add tests for subtyping of PolyT ~> PolyT
This tests for the unsoundness issue #8766.  It consists mainly
of the tests from #8767, with the expected results adapted to the
fact the issue isn't yet fixed.

As discussed at #8767, simply dropping the shortcut case and
relying on the general case for flows _ ~> PolyT isn't yet viable
because it causes some common code patterns to encounter a
different bug, producing large volumes of errors.  The plan is to
instead replace the shortcut with a different rule, from the
literature, which should be sound.

That rule fixes these unsoundness cases, and those are marked with
TODO comments.  It won't fix some other cases, where we reject valid
programs that would be natural to accept, because it still relies on
lining up the parameter lists against each other in order.

(Effectively the rule interprets the polymorphic types `<X: Sb>S`
and `<Y: Tb>T` as type *functions*, and asks whether the one
function lies below the other pointwise.  The general GenericT
machinery would instead take them as logical formulas headed by
the quantifier "∀", and seek to prove that the formula "∀X:Sb, S"
implies "∀Y:Tb, T".  That allows the lower tparam X to be inferred
as something other than Y, when that's what makes the proof work.)

We include tests for some of those cases too, marked as ideally OK
but known errors.