Fix unsound subtyping for PolyT ~> PolyT
Fixes #8766.

When checking a flow between polymorphic types PolyT ~> PolyT, we had
an attempted shortcut that wasn't sound, and also rejected some valid
flows.

The shortcut was added in 1aa5c63ba / D7454529 as an optimization to
avoid reaching the general case for flows _ ~> PolyT, which at the
time was the exponential `generate_tests`.  Happily new-generics means
that the general case is now efficient as well as correct.  So, simply
delete the code for the shortcut.

The test suite takes about 161s to run on my machine before this
change, and about 161s after it too.

The logic in question exists in two versions, one in subtyping_kit.ml
and one in flow_js.ml, in the unify function.  Both have the same
unsoundness; in fact both would accept each pair of types given as
examples in #8766.  But the unify version seems to also be dead code:
I tried replacing that case's body with an assert and ran the test
suite, and the assert never fired.  In any event, delete both.
Also fix and update some comments on this code.

In addition to the unsoundness in #8766, the tests demonstrate some
valid cases we previously rejected, including where the PolyT's have
different numbers of parameters.  Plus for completeness a couple of
cases we already got right, where one end of the flow is polymorphic
but the other isn't.