MakeExactT in annotation inference
Summary:
Ports support for `ExactT` and `MakeExactT` in annotation inference.
The code for this feature is limited enough, so that copying the logic was
much simpler than sharing the code like we did in earlier diffs. Besides,
the `cont` part of `Type.MakeExactT` does not have an equivalent in
annotation inference.

Changelog: [internal]

Reviewed By: jbrown215

Differential Revision: D29422806

fbshipit-source-id: abe29fd8b992e93584c9fa21f98acb8337e22335