LookupT in annotation inference
Summary:
`Annot_LookupT` is the annotation inference equivalent of `LookupT`, but simplified in
a few ways:
* The "kind" property is always `Strict` and so is omitted. The other available kinds
are not relevant in annotation inference, as they're related to unsealed objects.
* No `try_ts_on_failure`. There is no backtracking in annotation inference for now.

* There are a couple unlikely cases that are not support for now, and will be errors if
encountered:
  - Looking up "__proto__" on `ObjProtoT`, `FunProtoT`
  - Computed lookups on `NullT`, `ObjProtoT`, `FunProtoT`

Code sharing between Flow_js and here was really awkward, due the highly overloaded
nature of `LookupT`. Luckily, thanks to the above simplifications, the new code introduced
here is quite limited.

Changelog: [internal]

Reviewed By: jbrown215

Differential Revision: D29422811

fbshipit-source-id: 95cd03e43600ec7c07dcce4f7752ad7f4aeb1de8